// web-next/lib/types/feedback.ts
export interface FeedbackItem {
  id: string;
  uid?: string;
  codigoDoPersonal?: number;
  yourName?: string;
  comentarioDoAluno?: string;
  respostaDoProf?: string;
  estrela?: number;
  horaDeTermino?: Date;
  melhoria?: string;
  dificuldadeDoExercicio?: string;
  progresso?: string;
  imgUser?: string;
  nomeDoTreino?: string;
  obsInstrucao?: string;
  refPath?: string;
}